+++ macros/staging/remove_string_from_numeric.sql
@@ -6,25 +6,25 @@
 
 {% macro default__remove_string_from_numeric(column_name) %}
     
-    cast(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', '') as {{ dbt.type_numeric() }})
+    cast(nullif(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', ''), '') as {{ dbt.type_numeric() }})
 
 {% endmacro %}
 
 {% macro bigquery__remove_string_from_numeric(column_name) %}
 
-    cast(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), r'[^0-9.]*', '') as {{ dbt.type_numeric() }})
+    cast(nullif(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), r'[^0-9.]*', ''), '') as {{ dbt.type_numeric() }})
 
 {% endmacro %}
 
 {% macro postgres__remove_string_from_numeric(column_name) %}
 
-    cast(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', '', 'g') as {{ dbt.type_numeric() }})
+    cast(nullif(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', '', 'g'), '') as {{ dbt.type_numeric() }})
 
 {% endmacro %}
 
 {% macro redshift__remove_string_from_numeric(column_name) %}
     
-    cast(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', '') as {{ dbt.type_numeric() }})
+    cast(nullif(regexp_replace(cast({{ column_name }} as {{ dbt.type_string() }}), '[^0-9.]*', ''), '') as {{ dbt.type_numeric() }})
 
 {% endmacro %}